services/worker-image/requirements.txt
openai
structlog
neo4j
pillow
pydantic
pydantic-settings
celery[redis]
redis
sqlalchemy[asyncio]
asyncpg
python-dotenv
